+++ fabric/src/main/java/com/github/smallinger/copperagebackport/CopperAgeBackportFabric.java
@@ -139,6 +139,15 @@ public class CopperAgeBackportFabric implements ModInitializer {
             content.accept(ModItems.WAXED_OXIDIZED_COPPER_BUTTON_ITEM.get());
             // Copper Torch
             content.accept(ModItems.COPPER_TORCH_ITEM.get());
+            // Copper Lanterns
+            content.accept(ModItems.COPPER_LANTERN_ITEM.get());
+            content.accept(ModItems.EXPOSED_COPPER_LANTERN_ITEM.get());
+            content.accept(ModItems.WEATHERED_COPPER_LANTERN_ITEM.get());
+            content.accept(ModItems.OXIDIZED_COPPER_LANTERN_ITEM.get());
+            content.accept(ModItems.WAXED_COPPER_LANTERN_ITEM.get());
+            content.accept(ModItems.WAXED_EXPOSED_COPPER_LANTERN_ITEM.get());
+            content.accept(ModItems.WAXED_WEATHERED_COPPER_LANTERN_ITEM.get());
+            content.accept(ModItems.WAXED_OXIDIZED_COPPER_LANTERN_ITEM.get());
         });
         
         // Add copper tools to tools tab